- Use target_temp_step as in/decrements

+++ HAClimateDeskband/Models/SetTemperatureModel.cs
@@ -5,6 +5,6 @@ namespace HAClimateDeskband.Models
     public class SetTemperatureModel : EntityModel
     {
         [JsonProperty("temperature")]
-        public double Temperature { get; set; }
+        public decimal Temperature { get; set; }
     }
 }
